Remove obscure uses of shell redirections.
This patch makes the spec file slightly more friendly to non-shell
readers by changing "> foo" to "truncate -s 0 foo" and removes
the use of ">> foo". The use of ">> foo" is perhaps the most
interesting shell trick which is used to create a zero sized
debuginfocommon.filelist, but only if it doesn't already exist.
This allows the subsequent command to use debuginfocommon.filelist
without requiring it be wrapped in a check for
%{debuginfocommonarches}. That seems a little obtuse and certainly
confusing to the reader who expects such a check for anything
that is related to the debuginfo common package.