Fix failure load locale template.
The call to open_tmpl_archive was being passed a pointer to an
object allocated on the stack. The object on the stack is not
guaranteed to be initialized to zero so we need to minimally
initialize `fname' in the struct locarhandle to ensure that
open_tml_archive loads the default tempalte.

This error was seen while debugging glibc installs in a qemu
VM where it is more likely the stack pages were dirty. It has
not been reported on non-VM systems.