Build build-locale-archive statically again
build-locale-archive was switched to link dynamically in 538b3c08
without giving a proper reason for it.  The earlier static build was
wrong though, since it would happen against the installed glibc and
not the glibc being built.  The dynamic link was also similarly wrong,
more so because it would build against the built libc.so.6 and then
try to load the system libc.so.6.  This results in a failure in %post
in cases when the new build-locale-archive may have symbol references
that are not present in the old glibc.

There seem to be no good reason to run build-locale-archive with the
system libc.so.6, so the change is now reverted with a fixed up static
link that links against the build static libc.a.