Use scriptlet expansion.
We prefer scriptlet expansion (-e) over the use of rpm.expand in Lua.
The goal is to have a fully expanded Lua program with as little
dynamic processing as possible to simplify the operations carried
out in Lua.

The use of rpm.expand was only needed until COPR migrated from EL6
to Fedora, and with this complete we can remove the use of rpm.expand
in favour of scriptlet expansion.

Tested on Fedora Rawhide by verifying global setting of %_install_langs
macro changes installed locale-archive locales correctly.